Extract export utils to separate file
This allows unit tests to run without installing CKAN
