+++ ckanext/cfpb_extrafields/tests.py
@@ -3,7 +3,7 @@ from nose_parameterized import parameterized
 from nose.tools import assert_equal
 import mock
 import validators as v
-import ckanext.cfpb_extrafields.controllers.export as ec
+import ckanext.cfpb_extrafields.exportutils as eu
 
 
 class TestValidators(unittest.TestCase):
@@ -131,7 +131,7 @@ class TestExport(unittest.TestCase):
         ({"a": {"list_dict": [{"x": "y"}, {}]}}, {"a.list_dict": '[{"x": "y"}, {}]'}),
     ])
     def test_flatten(self, data, expected):
-        result = ec.flatten(data)
+        result = eu.flatten(data)
         assert_equal(result, expected)
 
     @parameterized.expand([
@@ -141,12 +141,12 @@ class TestExport(unittest.TestCase):
     ])
     def test_flatten_listseps(self, kwargs, expected):
         data = {"a": {"list_str": ["x", "y", "z"]}}
-        result = ec.flatten(data, **kwargs)
+        result = eu.flatten(data, **kwargs)
         assert_equal(result, expected)
     
     @parameterized.expand([
         ([{"a":1, "b": 2, "c": 3}], ["b", "a"], {"a": "A", "b": "bee"}, "bee,A\r\n2,1\r\n"),
     ])
     def test_to_csv(self, data, fields, fieldmap, expected):
-        result = ec.to_csv(data, fields, fieldmap)
+        result = eu.to_csv(data, fields, fieldmap)
         assert_equal(result, expected)
